VBA assignment part ii
This code did not included in the previous file, as it breaks as soon as I include them.  I think they should work, but I just not sure in which section i should insert these code to so that the For loop will work and populate the greatest% changes details.